MAINT: moving prints and plots to separated files (env_analysis)
