+++ rocketpy/plots/__init__.py
@@ -1,4 +1,5 @@
 from .compare import *
+from .environment_analysis_plots import _EnvironmentAnalysisPlots
 from .environment_plots import _EnvironmentPlots
 from .flight_plots import _FlightPlots
 from .rocket_plots import _RocketPlots
