+++ rocketpy/prints/__init__.py
@@ -1,5 +1,6 @@
 from .compare_prints import _ComparePrints
 from .environment_prints import _EnvironmentPrints
+from .environment_analysis_prints import _EnvironmentAnalysisPrints
 from .flight_prints import _FlightPrints
 from .parachute_prints import _ParachutePrints
 from .rocket_prints import _RocketPrints
